AnswerPage Javascript update
The script is mostly finished, however there is one bug i just cant fix.
The script will randomly assign one of the 4 tiles to be the correct
answer, The value of the correct answer is stored within a variable and
can be changed anytime. The bug is that the correct answer tile will
only work on the 4th tile, 1st, 2nd and 3rd tile share almost the
identical code yet they dont paint correctly.